1.0 (Release)
It turns out that there was a unit error in the Psi and Delta calculation that wasn't clear from the formulas. Most of the calculation is done in degrees, but for some strange reason, the alpha_1 needed to be in radians. Strange, but it's done. Other features have been requested and will be added in future updates.